👌 move contextHistory module in the tools folder
