+++ packages/rum-core/src/tools/contextHistory.spec.ts
