+++ packages/rum-core/src/tools/contextHistory.ts
